feat: moved job creation and checking into a dependency
